PVS and build warnings fixes

+++ src/xrGame/WeaponMagazinedWGrenade.cpp
@@ -747,7 +747,7 @@ void CWeaponMagazinedWGrenade::save(NET_Packet &output_packet)
 {
 	inherited::save								(output_packet);
 	save_data									(m_bGrenadeMode, output_packet);
-	save_data									(m_magazine2.size(), output_packet);
+	save_data									((u32)m_magazine2.size(), output_packet);
 
 }
 
@@ -765,7 +765,7 @@ void CWeaponMagazinedWGrenade::load(IReader &input_packet)
 	CCartridge					l_cartridge; 
 	l_cartridge.Load			(*m_ammoTypes2[m_ammoType2], u8(m_ammoType2));
 
-	while (sz > m_magazine2.size())
+	while (sz > (u32)m_magazine2.size())
 		m_magazine2.push_back(l_cartridge);
 }
